Auto get userIndex for logout
- Auto get userIndex for logout
- Code Format
- Add README
- Fix logging bug
- Add more logging record
